Removed protection settings from TopoAlign DLL
Removed several project files (TopoAlign2019.crproj to TopoAlign2025.crproj) from the "ConfuserProjects" project. These files contained rules and module paths related to the protection and renaming of namespaces and methods in the 'TopoAlign' DLL, which have now been removed. Additionally, the probe paths specifying the directories for DLL search have been removed. This suggests that the protection and obfuscation settings for the 'TopoAlign' DLL have been removed, and the DLL will no longer be searched for in the specified directories.