Updated product version and added support for new Revit and .NET versions
This commit includes several important updates. The `ProductCode` and `ProductVersion` in TopoAlign.aip have been updated to indicate a new product version (from `2.5.3` to `2.5.4`). The README.md file was updated to reflect the new versions of Revit and .NET that the software supports, changing the supported Revit versions from "2019 -- 2020" to "2025" and the .NET version from "4.7" to "8". A new badge was also added to display the total number of downloads from the GitHub repository. The TopoAlign.csproj file was updated to reflect the new software version. Lastly, a new instance of the `Microsoft.Xaml.Behaviors.DefaultTriggerAttribute` class was created in the AlignTopoView.xaml.cs file.