closes#13
Updated TextBox behaviors and bindings in AlignTopoView

Updated the regular expressions for `AllowableCharactersTextBoxBehavior` in the `TextBox` for "Divide edge distance" and "Vertical offset" to accept any number of digits before and after the decimal point, and allow for negative values in the latter. The `Text` binding for the `TextBox` for "Vertical offset" was also updated to trigger source updates on property changes, validate data errors, and notify on validation error. A `ToolTip` was added to guide users to enter a negative value for offsetting points above the topo. A redundant change in the namespace declaration for `behaviours` was also made.