"Removed build configurations for releases R19-R25 in TopoAlign.sln"
Removed several build configurations for releases R19 to R25 in both Debug and Release modes for the project with the ID {0331E696-1E32-44E0-BDA8-A1B1DBDDB4DC} in the TopoAlign.sln file. The configurations that were removed include the ActiveCfg and Build.0 settings for each release in both Debug and Release modes. No new code was added in this change.