Update build process and related files
This commit includes several updates to the build process and related files. The `.gitignore` file was updated to include `Bundle Testing`. The `build.schema.json` file was updated to include "Bundle" and "Publish" in the list of targets. The `TopoAlign.aip` file was updated to change the `ProductCode` and the source path of the `PackageContents.xml` file. The `README.md` file was updated to improve badge formatting. New files `Build.Bundle.cs` and `Build.Publish.cs` were added for new targets in the build process. The `Build.Compile.cs`, `Build.Installer.cs`, `Build.Sign.cs`, and `Build.cs` files were updated to improve the build process. A new file `PackageContents.cs` was added for handling the `ApplicationPackage` XML. The `_build.csproj` file was updated to change the target framework to `net8.0`.