Fix analytics.js path in multiple HTML files
Updated the `<script>` tag in the `<head>` section of six HTML files to correct the path of `analytics.js`. Changed the path from `./assets/analytics.js` to `../assets/analytics.js` to ensure proper resolution of the script file. This change addresses a directory-level issue and ensures the script is loaded correctly.

Affected files:
- CommandAlignFloor.html
- CommandAlignTopo.html
- CommandPointsAlongContours.html
- CommandPointsAtIntersection.html
- CommandPointsOnSurface.html
- CommandResetTopoRegion.html